Rename db/index.ts -> data/db.ts
