+++ src/useCases/AuthUseCases.ts
@@ -1,4 +1,4 @@
-import { db } from '@/db/index';
+import { db } from '@/data/db';
 import { BusinessError } from '@/shared/errors/BusinessError';
 import { Senha } from '@/valueObjects/Senha';
 import { nanoid } from 'nanoid';
